requirements.txt
python-telegram-bot
gspread
oauth2client
apscheduler
python-dotenv